Move ossl_sleep() to e_os.h and use it in apps
Fixes #15304
